requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch==2.8.0+cu128
torchvision==0.23.0+cu128
torchaudio==2.8.0+cu128

numpy==1.26.4
opencv-python==4.9.0.80
Pillow>=10.0
tqdm>=4.66
pyyaml>=6.0

ultralytics==8.3.205
onnxruntime-gpu==1.23.0 ; platform_system != "Darwin"
onnxruntime==1.23.0 ; platform_system == "Darwin"
open-clip-torch==3.2.0
